lb_listener_rules improvements
Life cycle rules:
We are planning to change the target group name of existing resources and
this is going to recreate the target group and attachment managed by this
module, and refresh the listener rule. We want to ensure the old target
group is only destroyed when the new one has been created and the listener
rule has been updated.

Target Group names:
Previous target groups created by the module ended up with names not very
user friendly, for instance `draft-fron-draft-finder-frontend`. We are using
these names in the configuration of Icinga checks. This format, originated
from the trimming of the `name` variable and the rule name, is making people
confused as it looks like a typo. We need to trim the target group name in
first place because it's limited to 32 characters.

The change here renames the variable to `target_group_prefix` to make it clear
how/where it's used. The description of the variable is improved. We are also
going to enforce the limit of this variable to 4 characters. With the use of
a null_resource resource, the plan/apply of the code will fail if the user
configures the variable with something longer than 4 chars, for instance:

```
Error: module.internal_lb_rules.null_resource.check_target_group_prefix_length:
: invalid or unknown key: ERROR: The length of target_group_prefix is longer than 4 characters
```

Hopefully this will force the user to use a short prefix that uniquely identifies
the target groups. We are expecting to use it with the "zone" of the Loadbalancer,
for instance: `int`, `ext`, `pub`, etc. The Target group tags can be used to search
for TGs instead of expecting to do all the identification from the name.