adds vendored plugin for docdb user provisioning
we need to provision documentdb users/roles, to do this we use a
third-party terraform provider so we don't have to script the logic of
create/destroy.

since we don't want to have to always install and keep in sync plugins
on all the developer devices and ci boxes that are used for terraform
runs, we vendor the binaries for linux/darwin into the repo so they can
be discovered.