Revert "Give the lb module target groups better names"
There seem to be issues with managing the target groups and related
load balancer listeners with Terraform. Terraform attempts to delete
the target groups (before creating new ones), but this fails as the
listeners depend on the target groups.

The Terraform configuration as it stands doesn't include dependencies
between the target groups and load balancer listeners, and this (or
the lack of the relationship) is recorded in the statefile.

Ideally, I think Terraform would create new listeners, with new target
groups, then delete the old listeners, then old target
groups. However, I think because the statefile doesn't include that
the listeners depend on the target groups, this doesn't happen. Also,
it looks like the statefile information is used, rather than the
Terraform configuration, as specifying the dependency explicitly in
the configuration seems to be ineffective.

This reverts commit 993bc7a184f79e01fec2c76f619ad889b5bde59b.